Update webhook example
