+++ kubernetes/apis/foo/v1alpha1/doc.go
