+++ kubernetes/apis/foo/v1alpha1/zz_generated.deepcopy.go
