+++ kubernetes/controller/main.go
@@ -20,7 +20,7 @@ import (
 	metricsserver "sigs.k8s.io/controller-runtime/pkg/metrics/server"
 	"sigs.k8s.io/controller-runtime/pkg/webhook"
 
-	foov1alpha1 "github.com/iawia002/pandora/kubernetes/apis/example/apis/foo/v1alpha1"
+	foov1alpha1 "github.com/iawia002/pandora/kubernetes/apis/foo/v1alpha1"
 	controllerruntime "github.com/iawia002/pandora/kubernetes/controller/controller-runtime"
 	samplecontroller "github.com/iawia002/pandora/kubernetes/controller/sample-controller"
 	controllerwebhook "github.com/iawia002/pandora/kubernetes/controller/webhook"
@@ -110,7 +110,9 @@ func run(config *rest.Config) error {
 	if err = (&foov1alpha1.Foo{}).SetupWebhookWithManager(mgr); err != nil {
 		return err
 	}
-	mgr.GetWebhookServer().Register("/mutate-v1-pod", &webhook.Admission{Handler: &controllerwebhook.PodAnnotator{Client: mgr.GetClient()}})
+	if err = controllerwebhook.SetupWebhookWithManager(mgr); err != nil {
+		return err
+	}
 
 	ctx := signals.SetupSignalHandler()
 	informer.Start(ctx.Done())
