+++ kubernetes/controller/webhook/pod.go
@@ -2,38 +2,75 @@ package webhook
 
 import (
 	"context"
-	"encoding/json"
-	"net/http"
+	"fmt"
 
 	corev1 "k8s.io/api/core/v1"
-	"sigs.k8s.io/controller-runtime/pkg/client"
+	"k8s.io/apimachinery/pkg/runtime"
+	ctrl "sigs.k8s.io/controller-runtime"
+	"sigs.k8s.io/controller-runtime/pkg/builder"
 	"sigs.k8s.io/controller-runtime/pkg/webhook/admission"
 )
 
-// PodAnnotator ...
-type PodAnnotator struct {
-	Client  client.Client
-	decoder *admission.Decoder
+// SetupWebhookWithManager ...
+func SetupWebhookWithManager(mgr ctrl.Manager) error {
+	return builder.WebhookManagedBy(mgr).
+		For(&corev1.Pod{}).
+		WithDefaulter(&podAnnotator{}).
+		WithValidator(&podValidator{}).
+		Complete()
 }
 
-// Handle ...
-func (a *PodAnnotator) Handle(_ context.Context, req admission.Request) admission.Response {
-	pod := &corev1.Pod{}
-	if err := a.decoder.Decode(req, pod); err != nil {
-		return admission.Errored(http.StatusBadRequest, err)
+// +kubebuilder:webhook:path=/mutate--v1-pod,mutating=true,failurePolicy=fail,sideEffects=None,groups="",resources=pods,verbs=create;update,versions=v1,name=mpod.kb.io,admissionReviewVersions=v1
+
+// podAnnotator annotates Pods
+type podAnnotator struct{}
+
+// Default ...
+func (a *podAnnotator) Default(_ context.Context, obj runtime.Object) error {
+	pod, ok := obj.(*corev1.Pod)
+	if !ok {
+		return fmt.Errorf("expected a Pod but got a %T", obj)
 	}
 
-	// mutate the fields in pod
-
-	marshaledPod, err := json.Marshal(pod)
-	if err != nil {
-		return admission.Errored(http.StatusInternalServerError, err)
+	if pod.Annotations == nil {
+		pod.Annotations = map[string]string{}
 	}
-	return admission.PatchResponseFromRaw(req.Object.Raw, marshaledPod)
-}
-
-// InjectDecoder ...
-func (a *PodAnnotator) InjectDecoder(d *admission.Decoder) error {
-	a.decoder = d
+	pod.Annotations["example-mutating-admission-webhook"] = "foo"
 	return nil
+}
+
+// +kubebuilder:webhook:path=/validate--v1-pod,mutating=false,failurePolicy=fail,sideEffects=None,groups="",resources=pods,verbs=create;update,versions=v1,name=vpod.kb.io,admissionReviewVersions=v1
+
+// podValidator validates Pods
+type podValidator struct{}
+
+// validate admits a pod if a specific annotation exists.
+func (v *podValidator) validate(_ context.Context, obj runtime.Object) (admission.Warnings, error) {
+	pod, ok := obj.(*corev1.Pod)
+	if !ok {
+		return nil, fmt.Errorf("expected a Pod but got a %T", obj)
+	}
+
+	key := "example-mutating-admission-webhook"
+	anno, found := pod.Annotations[key]
+	if !found {
+		return nil, fmt.Errorf("missing annotation %s", key)
+	}
+	if anno != "foo" {
+		return nil, fmt.Errorf("annotation %s did not have value %q", key, "foo")
+	}
+
+	return nil, nil
+}
+
+func (v *podValidator) ValidateCreate(ctx context.Context, obj runtime.Object) (admission.Warnings, error) {
+	return v.validate(ctx, obj)
+}
+
+func (v *podValidator) ValidateUpdate(ctx context.Context, _, newObj runtime.Object) (admission.Warnings, error) {
+	return v.validate(ctx, newObj)
+}
+
+func (v *podValidator) ValidateDelete(ctx context.Context, obj runtime.Object) (admission.Warnings, error) {
+	return v.validate(ctx, obj)
 }
