+++ kubernetes/crds/foo.example.io_foos.yaml
