+++ kubernetes/hack/update-all.sh
