+++ kubernetes/scheme/scheme.go
@@ -4,7 +4,7 @@ import (
 	"k8s.io/apimachinery/pkg/runtime"
 	clientgoscheme "k8s.io/client-go/kubernetes/scheme"
 
-	foov1alpha1 "github.com/iawia002/pandora/kubernetes/apis/example/apis/foo/v1alpha1"
+	foov1alpha1 "github.com/iawia002/pandora/kubernetes/apis/foo/v1alpha1"
 )
 
 // Scheme contains all types of custom Scheme and kubernetes client-go Scheme.
